+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/SdkSynchronizer.kt
@@ -669,7 +669,28 @@ class SdkSynchronizer private constructor(
             account
         )
 
-    @Throws(TransactionEncoderException::class)
+    /**
+     * Creates a proposal for fulfilling a payment ZIP-321 URI
+     *
+     * @param account the account from which to transfer funds.
+     * @param uri a ZIP-321 compliant payment URI String
+     *
+     * @throws TransactionEncoderException.ProposalFromUriException in case the proposal creation failed
+     *
+     * @return the proposal or an exception
+     */
+    @Throws(TransactionEncoderException.ProposalFromUriException::class)
+    override suspend fun proposeFulfillingPaymentUri(
+        account: Account,
+        uri: String
+    ): Proposal = txManager.proposeTransferFromUri(account, uri)
+
+    /**
+     * @throws TransactionEncoderException.ProposalFromParametersException in case the proposal creation failed
+     *
+     * @return the proposal or an exception
+     */
+    @Throws(TransactionEncoderException.ProposalFromParametersException::class)
     override suspend fun proposeTransfer(
         account: Account,
         recipient: String,
@@ -677,7 +698,12 @@ class SdkSynchronizer private constructor(
         memo: String
     ): Proposal = txManager.proposeTransfer(account, recipient, amount, memo)
 
-    @Throws(TransactionEncoderException::class)
+    /**
+     * @throws TransactionEncoderException.ProposalShieldingException in case the proposal creation failed
+     *
+     * @return the proposal or an exception
+     */
+    @Throws(TransactionEncoderException.ProposalShieldingException::class)
     override suspend fun proposeShielding(
         account: Account,
         shieldingThreshold: Zatoshi,
@@ -685,7 +711,10 @@ class SdkSynchronizer private constructor(
         transparentReceiver: String?
     ): Proposal? = txManager.proposeShielding(account, shieldingThreshold, memo, transparentReceiver)
 
-    @Throws(TransactionEncoderException::class)
+    @Throws(
+        TransactionEncoderException.TransactionNotCreatedException::class,
+        TransactionEncoderException.TransactionNotFoundException::class
+    )
     override suspend fun createProposedTransactions(
         proposal: Proposal,
         usk: UnifiedSpendingKey
@@ -831,7 +860,7 @@ class SdkSynchronizer private constructor(
             ) {
                 is Response.Success -> {
                     Twig.info { "Chain tip for validate consensus branch action fetched: ${response.result.value}" }
-                    runCatching { response.result.toBlockHeight(network) }.getOrNull()
+                    runCatching { response.result.toBlockHeight() }.getOrNull()
                 }
 
                 is Response.Failure -> {
@@ -890,7 +919,7 @@ class SdkSynchronizer private constructor(
 
             // Check sapling activation height
             runCatching {
-                val remoteSaplingActivationHeight = remoteInfo.saplingActivationHeightUnsafe.toBlockHeight(network)
+                val remoteSaplingActivationHeight = remoteInfo.saplingActivationHeightUnsafe.toBlockHeight()
                 if (network.saplingActivationHeight != remoteSaplingActivationHeight) {
                     return ServerValidation.InValid(
                         CompactBlockProcessorException.MismatchedSaplingActivationHeight(
@@ -906,7 +935,7 @@ class SdkSynchronizer private constructor(
             val currentChainTip =
                 when (val response = lightWalletClient.getLatestBlockHeight()) {
                     is Response.Success -> {
-                        runCatching { response.result.toBlockHeight(network) }.getOrElse {
+                        runCatching { response.result.toBlockHeight() }.getOrElse {
                             return ServerValidation.InValid(it)
                         }
                     }
@@ -986,7 +1015,6 @@ internal object DefaultSynchronizerFactory {
         context: Context,
         rustBackend: TypesafeBackend,
         databaseFile: File,
-        zcashNetwork: ZcashNetwork,
         checkpoint: Checkpoint,
         seed: ByteArray?,
         numberOfAccounts: Int,
@@ -997,7 +1025,6 @@ internal object DefaultSynchronizerFactory {
                 context,
                 rustBackend,
                 databaseFile,
-                zcashNetwork,
                 checkpoint,
                 seed,
                 numberOfAccounts,