+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/Synchronizer.kt
@@ -214,6 +214,19 @@ interface Synchronizer {
     ): Proposal
 
     /**
+     * Creates a proposal for fulfilling a payment ZIP-321 URI
+     *
+     * @param account the account from which to transfer funds.
+     * @param uri a ZIP-321 compliant payment URI String
+     *
+     * @return the proposal or an exception
+     */
+    suspend fun proposeFulfillingPaymentUri(
+        account: Account,
+        uri: String
+    ): Proposal
+
+    /**
      * Creates a proposal for shielding any transparent funds received by the given account.
      *
      * @param account the account for which to shield funds.
@@ -650,7 +663,7 @@ interface Synchronizer {
                         when (val response = downloader.getLatestBlockHeight()) {
                             is Response.Success -> {
                                 Twig.info { "Chain tip for recovery until param fetched: ${response.result.value}" }
-                                runCatching { response.result.toBlockHeight(zcashNetwork) }.getOrNull()
+                                runCatching { response.result.toBlockHeight() }.getOrNull()
                             }
                             is Response.Failure -> {
                                 Twig.error {
@@ -670,7 +683,6 @@ interface Synchronizer {
                     context = applicationContext,
                     rustBackend = backend,
                     databaseFile = coordinator.dataDbFile(zcashNetwork, alias),
-                    zcashNetwork = zcashNetwork,
                     checkpoint = loadedCheckpoint,
                     seed = seed,
                     numberOfAccounts = Derivation.DEFAULT_NUMBER_OF_ACCOUNTS,