+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/TypesafeBackendImpl.kt
@@ -40,6 +40,17 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
         )
     }
 
+    override suspend fun proposeTransferFromUri(
+        account: Account,
+        uri: String
+    ): Proposal =
+        Proposal.fromUnsafe(
+            backend.proposeTransferFromUri(
+                account.value,
+                uri
+            )
+        )
+
     @Suppress("LongParameterList")
     override suspend fun proposeTransfer(
         account: Account,
@@ -99,10 +110,7 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
     }
 
     override suspend fun getNearestRewindHeight(height: BlockHeight): BlockHeight {
-        return BlockHeight.new(
-            ZcashNetwork.from(backend.networkId),
-            backend.getNearestRewindHeight(height.value)
-        )
+        return BlockHeight.new(backend.getNearestRewindHeight(height.value))
     }
 
     override suspend fun rewindToHeight(height: BlockHeight) {
@@ -111,10 +119,7 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
 
     override suspend fun getLatestCacheHeight(): BlockHeight? {
         return backend.getLatestCacheHeight()?.let {
-            BlockHeight.new(
-                ZcashNetwork.from(backend.networkId),
-                it
-            )
+            BlockHeight.new(it)
         }
     }
 
@@ -204,10 +209,7 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
     override suspend fun getFullyScannedHeight(): BlockHeight? {
         return runCatching {
             backend.getFullyScannedHeight()?.let {
-                BlockHeight.new(
-                    ZcashNetwork.from(backend.networkId),
-                    it
-                )
+                BlockHeight.new(it)
             }
         }.onFailure {
             Twig.warn(it) { "Currently unable to get fully scanned height" }
@@ -217,10 +219,7 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
     override suspend fun getMaxScannedHeight(): BlockHeight? {
         return runCatching {
             backend.getMaxScannedHeight()?.let {
-                BlockHeight.new(
-                    ZcashNetwork.from(backend.networkId),
-                    it
-                )
+                BlockHeight.new(it)
             }
         }.onFailure {
             Twig.warn(it) { "Currently unable to get max scanned height" }
@@ -231,14 +230,11 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
         fromHeight: BlockHeight,
         fromState: TreeState,
         limit: Long
-    ): ScanSummary = ScanSummary.new(backend.scanBlocks(fromHeight.value, fromState.encoded, limit), network)
+    ): ScanSummary = ScanSummary.new(backend.scanBlocks(fromHeight.value, fromState.encoded, limit))
 
     override suspend fun transactionDataRequests(): List<TransactionDataRequest> =
         backend.transactionDataRequests().map { jniRequest ->
-            TransactionDataRequest.new(
-                jniRequest,
-                network
-            )
+            TransactionDataRequest.new(jniRequest)
         }
 
     override suspend fun getWalletSummary(): WalletSummary? =
@@ -248,10 +244,7 @@ internal class TypesafeBackendImpl(private val backend: Backend) : TypesafeBacke
 
     override suspend fun suggestScanRanges(): List<ScanRange> =
         backend.suggestScanRanges().map { jniScanRange ->
-            ScanRange.new(
-                jniScanRange,
-                network
-            )
+            ScanRange.new(jniScanRange)
         }
 
     override suspend fun decryptAndStoreTransaction(