+++ sdk-lib/src/test/java/cash/z/ecc/android/sdk/model/BlockHeightTest.kt
@@ -6,34 +6,24 @@ import kotlin.test.assertFailsWith
 
 class BlockHeightTest {
     @Test
-    fun new_mainnet_fails_below_sapling_activation_height() {
-        assertFailsWith(IllegalArgumentException::class) {
-            BlockHeight.new(
-                ZcashNetwork.Mainnet,
-                ZcashNetwork.Mainnet.saplingActivationHeight.value - 1
-            )
-        }
-    }
-
-    @Test
     fun new_mainnet_succeeds_at_sapling_activation_height() {
-        BlockHeight.new(ZcashNetwork.Mainnet, ZcashNetwork.Mainnet.saplingActivationHeight.value)
+        BlockHeight.new(ZcashNetwork.Mainnet.saplingActivationHeight.value)
     }
 
     @Test
     fun new_mainnet_succeeds_above_sapling_activation_height() {
-        BlockHeight.new(ZcashNetwork.Mainnet, ZcashNetwork.Mainnet.saplingActivationHeight.value + 10_000)
+        BlockHeight.new(ZcashNetwork.Mainnet.saplingActivationHeight.value + 10_000)
     }
 
     @Test
     fun new_mainnet_succeeds_at_max_value() {
-        BlockHeight.new(ZcashNetwork.Mainnet, UInt.MAX_VALUE.toLong())
+        BlockHeight.new(UInt.MAX_VALUE.toLong())
     }
 
     @Test
     fun new_fails_above_max_value() {
         assertFailsWith(IllegalArgumentException::class) {
-            BlockHeight.new(ZcashNetwork.Mainnet, UInt.MAX_VALUE.toLong() + 1)
+            BlockHeight.new(UInt.MAX_VALUE.toLong() + 1)
         }
     }
 
@@ -65,11 +55,10 @@ class BlockHeightTest {
     fun subtraction_of_block_height_succeeds() {
         val one =
             BlockHeight.new(
-                ZcashNetwork.Mainnet,
                 ZcashNetwork.Mainnet.saplingActivationHeight.value +
                     ZcashNetwork.Mainnet.saplingActivationHeight.value
             )
-        val two = BlockHeight.new(ZcashNetwork.Mainnet, ZcashNetwork.Mainnet.saplingActivationHeight.value)
+        val two = BlockHeight.new(ZcashNetwork.Mainnet.saplingActivationHeight.value)
 
         assertEquals(ZcashNetwork.Mainnet.saplingActivationHeight.value, one - two)
     }