+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- The new `Synchronizer.proposeFulfillingPaymentUri` API has been added. It enables constructing Proposal object from 
+  the given ZIP-321 Uri, and then creating transactions from it.
+
+### Changed
+- `Synchronizer.proposeTransfer` throws `TransactionEncoderException.ProposalFromParametersException`
+- `Synchronizer.proposeShielding` throws `TransactionEncoderException.ProposalShieldingException`
+- `Synchronizer.createProposedTransactions` throws `TransactionEncoderException.TransactionNotCreatedException` and `TransactionEncoderException.TransactionNotFoundException`
 - `LightWalletClient` now implements `Closeable` and is thus correctly cleaned up in `SdkSynchronizer` and
   `FastestServerFetcher` after it's used